+++ fulltext/backends/__odt.py
@@ -8,7 +8,7 @@ from six import StringIO
 
 from fulltext import BaseBackend
 from fulltext.util import run, assert_cmd_exists
-from fulltext import is_file_path
+from fulltext.util import is_file_path
 
 
 def qn(ns):
